% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stan_glm.R
\name{stan_glm}
\alias{stan_glm}
\title{stan_glm}
\usage{
stan_glm(formula, data, family = "gaussian", prior = "dnorm(0,10)",
  loc_scale_transform = TRUE, file_name = NULL, run_stan = FALSE, ...)
}
\arguments{
\item{formula}{Formula (using lmer syntax).}

\item{data}{A \code{data.frame} containing the variables in the model.}

\item{family}{Family.}

\item{prior}{Prior distribution for regression coefficients.}

\item{loc_scale_transform}{Logical. Use non-centered parameterizations?}

\item{file_name}{Full path for .stan file to be created.}

\item{run_stan}{Logical. Fit the model using \pkg{rstan} in addition
to creating the .stan file?}

\item{...}{If \code{run_stan = TRUE}, optional arguments to pass to
\code{stan} (e.g. \code{iter}, \code{chains}, etc.).
See \code{\link[rstan]{stan}} (\pkg{rstan}).}
}
\description{
stan_glm
}

