% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stan_lmer.R
\name{stan_lmer}
\alias{stan_lmer}
\title{stan_lmer}
\usage{
stan_lmer(fl, data, family = gaussian, prefix = c("b_", "v_"),
  default_prior = "dnorm(0,10)", matt_trick = TRUE, file_name = NULL, ...)
}
\description{
stan_lmer
}

