% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stanRegression.R
\docType{package}
\name{stanRegression}
\alias{stanRegression}
\alias{stanRegression-package}
\title{stanRegression}
\description{
Fit regression models with Stan using familiar lm, glm,
and lmer syntax.
}

