% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stan_lmer.R
\name{stan_lmer}
\alias{stan_lmer}
\title{stan_lmer}
\usage{
stan_lmer(fl, data, family = gaussian, prefix = c("b_", "v_"),
  default_prior = "dnorm(0,10)", matt_trick = TRUE, file_name = NULL,
  run_stan = FALSE, ...)
}
\arguments{
\item{fl}{Formula.}

\item{data}{Data. Must be a \code{data.frame}.}

\item{family}{}

\item{prefix}{}

\item{default_prior}{Default prior for 'fixed effects'.}

\item{matt_trick}{Logical. Use NCP?}

\item{file_name}{Full path for .stan file to be created.}

\item{run_stan}{Logical. Should the model be run in RStan?}
}
\description{
stan_lmer
}

